Rename CreationUiState and navigation objects
