+++ app/src/main/java/com/example/wateringreminder/compose/navigation/BottomNavGraph.kt
@@ -1,11 +1,10 @@
-package com.example.wateringreminder
+package com.example.wateringreminder.compose.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.example.wateringreminder.articlelist.MyPlantsScreen
-import com.example.wateringreminder.compose.navigation.BottomBarScreen
 import com.example.wateringreminder.watering.WateringScreen
 
 
@@ -13,14 +12,14 @@ import com.example.wateringreminder.watering.WateringScreen
 fun BottomNavGraph(navController: NavHostController) {
     NavHost(
         navController = navController,
-        startDestination = BottomBarScreen.Plants.route
+        startDestination = BottomBarScreen.Watering.route
     ) {
-        composable(route = BottomBarScreen.Plants.route) {
+        composable(route = BottomBarScreen.Watering.route) {
             WateringScreen(
                 onNavigateToPlantCreator = { navController.navigate("plantCreator")}
             )
         }
-        composable(route = BottomBarScreen.Articles.route) {
+        composable(route = BottomBarScreen.MyPlants.route) {
             MyPlantsScreen()
         }
         plantNavGraph(navController)
