+++ app/src/main/java/com/example/wateringreminder/plantcreator/PlantCreationViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.data_source.Plant
 import com.example.data_source.PlantRepository
-import com.example.wateringreminder.watering.CreationUiState
+import com.example.wateringreminder.watering.WateringUiState
 import com.example.wateringreminder.EmptyTextFieldValidator
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -19,8 +19,8 @@ class PlantCreationViewModel(
     private val plantRepository: PlantRepository
 ) : ViewModel() {
 
-    private val _uiState = MutableStateFlow(CreationUiState())
-    val uiState: StateFlow<CreationUiState> = _uiState.asStateFlow()
+    private val _uiState = MutableStateFlow(WateringUiState())
+    val uiState: StateFlow<WateringUiState> = _uiState.asStateFlow()
 
     private var plantName by mutableStateOf("")
     private var location by mutableStateOf("")
