+++ app/src/main/java/com/example/wateringreminder/watering/WateringUiState.kt
@@ -2,7 +2,7 @@ package com.example.wateringreminder.watering
 
 import com.example.wateringreminder.R
 
-data class CreationUiState(
+data class WateringUiState(
     var showNameError: Boolean = false,
     var errorNameMsg: Int = R.string.empty_string,
     var plantName: String = "",
